Rewrite to only serve APIs
